gamedata implemented; botdataxmlhandler updated; some int swapped with size_t; DynObj removed

+++ GameModule/src/BotManager.cpp
@@ -70,7 +70,7 @@ void BotManager::receiveBroadcast(int fromID, BroadcastMessage msg, int dataSize
 				// copying data, can't let bots delete it
 				int *dataCopy;
 				dataCopy = new int[dataSize];
-				for (int i = 0; i < dataSize; ++i)
+				for (size_t i = 0; i < dataSize; ++i)
 					dataCopy[i] = data[i];
  				this->bot->roundWinners(dataSize, dataCopy);
 				delete [] dataCopy;
