+++ GameModule/src/Table.h
@@ -10,7 +10,7 @@
 class Table
 {
 private:
-	int numOfBots;
+	size_t numOfBots;
 	BotInfo **bots;
 	int numOfCards;
 	Card* cards[5];
@@ -24,13 +24,13 @@ public:
 		this->bots = new BotInfo*[numOfBots];
 
 		// clear bots array
-		for (int i = 0; i < numOfBots; ++i)
+		for (size_t i = 0; i < numOfBots; ++i)
 		{
 			this->bots[i] = nullptr;
 		}
 
 		// clear cards array
-		for (int i = 0; i < 5; ++i)
+		for (size_t i = 0; i < 5; ++i)
 		{
 			this->cards[i] = nullptr;
 		}
